newsletter-editor/requirements.txt
feedparser>=6.0
httpx>=0.27
beautifulsoup4>=4.12
tweepy>=4.14
PyJWT>=2.8
pyyaml>=6.0
